% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxCox.R
\name{boxCox}
\alias{boxCox}
\title{Box-Cox transform}
\usage{
boxCox(x, lambda)
}
\arguments{
\item{x:}{input list}

\item{lambda:}{transformation parameter}
}
\value{
the Box-Cox tranformed data
}
\description{
Box-Cox transform
}
